systemd/rpm - provision systemd service, do proper upgrade
Provision a systemd service that starts the vocalsalad consolescript,
the main function in server.py. Moreover get the RPM to detect if it
is currently in an install/uninstall or an upgrade situation.
Upgrading.

-	If we're just installing from scratch copy the files, create the
	service, and start it.
- 	If we're just uninstalling from scratch delete the files,
	stop/disable the service.
-	If we're upgrading then we need to stop the service, copy the
	files, then restart the service. Very naughty if we just left the
	service running and copied/deleted files under its feet! And
	even worse if we forgot to restart the service.
-	If we're upgrading then the post-install step must not delete
	any files (!!).

Note that when installing RPMs reading [1] is a good tutorial on the
steps taking during upgrades/installs.

[1] https://www.ibm.com/developerworks/library/l-rpm2/

+++ Makefile
@@ -1,5 +1,5 @@
 VERSION = 0.0.1
-ITERATION = 2
+ITERATION = 5
 
 MAINTAINER = asim.ihsan@gmail.com
 URL = https://github.com/asimihsan/vocalsalad
@@ -8,17 +8,73 @@ PYTHON_BIN = python2.7
 PROJECT = vocalsalad
 ROOT = /opt/$(PROJECT)
 
-define POST_INSTALL_SH
+define BEFORE_INSTALL_SH
 #!/usr/bin/env bash
-ln -s $(ROOT)/bin/$(PROJECT) /usr/local/bin/$(PROJECT)
+set -e
 endef
-export POST_INSTALL_SH
+export BEFORE_INSTALL_SH
 
-define POST_UNINSTALL_SH
+define AFTER_INSTALL_SH
 #!/usr/bin/env bash
-unlink /usr/local/bin/$(PROJECT)
+set -e
+
+if [ "$$1" = "1" ]; then
+	# Perform tasks to prepare for the initial installation (no previous
+	# installation is present)
+	echo "AFTER_INSTALL: This is an initial installation. Create/copy "
+	echo "files, enable/start the service."
+	ln -s $(ROOT)/bin/$(PROJECT) /usr/local/bin/$(PROJECT)
+	if [ ! -f /etc/systemd/system/$(PROJECT).service ]; then
+		cp $(ROOT)/infrastructure/fedora/$(PROJECT).service \
+			/etc/systemd/system/$(PROJECT).service
+	fi
+	systemctl daemon-reload
+	systemctl enable $(PROJECT).service
+	systemctl start $(PROJECT).service
+elif [ "$$1" = "2" ]; then
+	# Perform whatever maintenance must occur before the upgrade begins
+	echo "AFTER_INSTALL: This is an upgrade. Just restart the service."
+	systemctl restart $(PROJECT).service
+fi
 endef
-export POST_UNINSTALL_SH
+export AFTER_INSTALL_SH
+
+define BEFORE_REMOVE_SH
+#!/usr/bin/env bash
+set -e
+
+if [ "$$1" = "0" ]; then
+	# Perform tasks to prepare for the uninstallation (we are not
+	# upgrading so brutally clean everything up).
+	echo "BEFORE_REMOVE: This is a final uninstallation. Stop/disable "
+	echo "service, remove service file."
+	if [ -f /etc/systemd/system/$(PROJECT).service ]; then
+		systemctl stop $(PROJECT).service
+		systemctl disable $(PROJECT).service
+		rm -f /etc/systemd/system/$(PROJECT).service
+	fi;
+elif [ "$$1" = "1" ]; then
+	# Perform whatever maintenance must occur before the upgrade ends
+	echo "BEFORE_REMOVE: This is an upgrade. Do nothing."
+fi
+endef
+export BEFORE_REMOVE_SH
+
+define AFTER_REMOVE_SH
+#!/usr/bin/env bash
+set -e
+
+if [ "$$1" = "0" ]; then
+	# Perform tasks to prepare for the uninstallation (we are not
+	# upgrading so brutally clean everything up).
+	echo "AFTER_REMOVE: This is a final uninstallation. Remove files."
+	unlink /usr/local/bin/$(PROJECT)
+elif [ "$$1" = "1" ]; then
+	# Perform whatever maintenance must occur before the upgrade ends
+	echo "AFTER_REMOVE: This is an upgrade. Do nothing."
+fi
+endef
+export AFTER_REMOVE_SH
 
 rpm:
 	sudo chown vagrant:vagrant /opt
@@ -27,17 +83,26 @@ rpm:
 		--system-site-packages $(ROOT)
 	source "$(ROOT)/bin/activate"; python setup.py install
 	virtualenv --relocatable $(ROOT)
-	echo "$$POST_INSTALL_SH" > /tmp/post-install.sh
-	echo "$$POST_UNINSTALL_SH" > /tmp/post-uninstall.sh
+	cp -r /vagrant/$(PROJECT)/infrastructure $(ROOT)
+
+	echo "$$BEFORE_INSTALL_SH" > /tmp/before-install.sh
+	echo "$$AFTER_INSTALL_SH" > /tmp/after-install.sh
+	echo "$$BEFORE_REMOVE_SH" > /tmp/before-remove.sh
+	echo "$$AFTER_REMOVE_SH" > /tmp/after-remove.sh
 	rm -f *.rpm
 	source /etc/os-release; fpm -s dir -t rpm -n '$(PROJECT)' \
 		--package-name-suffix "$$ID$$VERSION_ID" \
 		--version $(VERSION) --iteration $(ITERATION) --epoch 0 \
 		-C $(ROOT) --python-bin $(PYTHON_BIN) \
-		--directories=$(ROOT) --post-install /tmp/post-install.sh \
-	    --post-uninstall /tmp/post-uninstall.sh --architecture native \
-	    --depends python --maintainer $(MAINTAINER) --url $(URL) \
+		--directories=$(ROOT) \
+		--before-install /tmp/before-install.sh \
+		--after-install /tmp/after-install.sh \
+		--before-remove /tmp/before-remove.sh \
+	    --after-remove /tmp/after-remove.sh \
+	    --architecture native --depends python \
+	    --maintainer $(MAINTAINER) --url $(URL) \
 	    $(ROOT)
 	mv *.rpm /vagrant/package-dist/
 	rm -rf $(ROOT)
-	rm -f /tmp/post-install.sh /tmp/post-uninstall.sh
+	rm -f /tmp/before-install.sh /tmp/after-install.sh \
+		/tmp/before-remove.sh /tmp/after-remove.sh
